+++ src/Entities/PayFac/DeviceDetails.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace GlobalPayments\Api\Entities\PayFac;
+
+class DeviceDetails
+{
+    public $quantity;
+    public $timezone;
+}